refactor: prompt beta to prod
